Fix the dependencies of functions on cosmowrapper
It was on inputs, it has to be on cosmology.py
